+++ com.mathewhdyt.audiomanager/Runtime/AudioManager/Locator/NullAudioManager.cs
@@ -43,6 +43,10 @@ namespace AudioManager.Locator {
             return AudioError.NOT_INITIALIZED;
         }
 
+        public AudioError DeregisterChild(string name, ChildType child) {
+            return AudioError.NOT_INITIALIZED;
+        }
+
         public AudioError PlayDelayed(string name, float delay, ChildType child) {
             return AudioError.NOT_INITIALIZED;
         }